src/test/java/com/simi/studies/csjava/datastructure/lifo/FixedCapacityStackTest.java
package com.simi.studies.csjava.datastructure.lifo;

import com.simi.studies.csjava.datastructure.CapacityExceededException;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

class FixedCapacityStackTest {

  @Test
  public void givenStackWithoutElements_whenCheckingEmptiness_thenShouldBeTrue() {
    final var emptyStack = new FixedCapacityStack<String>(0);
    Assertions.assertTrue(emptyStack.isEmpty());
  }

  @Test
  public void givenStackWithOneElement_whenCheckingEmptiness_thenShouldBeFalse() {
    final var fixedCapacity = new FixedCapacityStack<String>(1);
    fixedCapacity.push("Test");
    Assertions.assertFalse(fixedCapacity.isEmpty());
  }

  @Test
  public void givenStackWithMultipleElements_whenCheckingSize_thenShouldTwo() {
    final var fixedCapacity = new FixedCapacityStack<String>(2);
    fixedCapacity.push("Test1");
    fixedCapacity.push("Test2");
    Assertions.assertEquals(2, fixedCapacity.size());
  }

  @Test
  public void givenStackWithMultipleElements_whenExceedingCapacity_thenThrowException() {
    final var fixedCapacity = new FixedCapacityStack<String>(2);
    fixedCapacity.push("Test1");
    fixedCapacity.push("Test2");
    Assertions.assertThrows(CapacityExceededException.class, () -> fixedCapacity.push("Test3"));
  }

}